+++ src/index.js
@@ -1,6 +1,6 @@
 /**
  * This is the main package file.
  */
-export default async function expensive2() {
-  console.log('expensive2 called')
+export default async function expensive() {
+  console.log('expensive called')
 }