+++ test/spec/index.js
@@ -1,15 +1,15 @@
 import { equal } from 'zoroaster/assert'
 import Context from '../context'
-import expensive2 from '../../src'
+import expensive from '../../src'
 
 /** @type {Object.<string, (c: Context)>} */
 const T = {
   context: Context,
   'is a function'() {
-    equal(typeof expensive2, 'function')
+    equal(typeof expensive, 'function')
   },
   async 'calls package without error'() {
-    await expensive2()
+    await expensive()
   },
   async 'calls test context method'({ example }) {
     await example()